feat: organize source and test files into respective directories; update CMakeLists.txt and test scripts
